use all the logcheck-* files from violations.ignore.d!
but now we have a potential race:
if $grepfile = logcheck and a pattern of it matches.
we would run 2x times the logcheck-* rules.

i prefer to remove the stuff below
# Now ignore all entries from the logcheck-<package> files
will do it next days if no one waves hand. *)

bonus: less grep values to check ;)

i'm not shure if the stuff below debug "Cleaning logcheck"
is needed for now I leave it.

*) readded file restriction check that was accidentaly removed,
   by a patch of mine for version 1.2.17.