Enable alternate temporary directory to be set in the configuration file
On servers where the logfiles are very large and grow quickly, the
logcheck processes may run into space issues in /tmp resulting in the
email such as the following to be sent:

cp: writing `/tmp/logcheck.y12449/checked': No space left on device
/usr/sbin/logcheck: line 161: cannot create temp file for here document: No space left on device
mail: /tmp/mail.RsXXXXpc2eAx: No space left on device
Null message body; hope that's ok

Followed by an empty email.

This is not good because of a few reasons:

1. /tmp fills up temporarily, other problems can result from this
2. You loose logcheck for this run

Although logcheck falls back to /var/tmp if it cannot create a directory in /tmp,
it doesn't allow the admin to set a temporary directory other than /tmp if this
condition presents itself.

The solution is simple, the attached patch adds a new configuration to
/etc/logcheck/logcheck.conf for a TMP variable, sets it to be default
what logcheck uses now (/tmp) and alters the logcheck executable to honor this
variable when set.

Signed-off-by: martin f. krafft <madduck@madduck.net>