Replaced bashisms with POSIX equivalents
Since logcheck doesn't require any bash-specific feature, there's no
harm in making it as POSIX-y as possible.  This was actually requested
on the BTS, so why not?

However, I'm not changing the shebang to /bin/sh, since it is far from
certain that the script is now POSIX compliant.  (Besides, what would be
the point?  Any speed gain will be dwarfed by egrep anyway.)