Adding support for removing conffiles in logcheck-database's preinst
This basically copies rm_conffile() from the Debian wiki, to deal with
the fact that dpkg does not delete conffiles after they have been
removed from a package.

This version of rm_conffile() makes sure that the conffile still belongs
to our package, since it is possible for it to have migrated to another
package through a Replaces relationship.  (See bug #532484 for an
example.)

Thanks to this check, we can also avoid bothering with version
comparisions; calling rm_conffile() on subsequent upgrades will simply
do nothing.  (Of course, we still avoid calling it uselessly on fresh
installs.)