Support removing even older conffiles in logcheck-database's preinst
Many conffiles have come and gone throughout logcheck's history, and due
to the fact that dpkg did not remove them, these obsolete files are
still left rotting on many installations.  (See bug #453519 for an
example.)

This attempts to delete those files on install/upgrade, after making
sure they have not been "adopted" by another package in the meantime.
If a file is deemed to have been modified, it will be backed up and
subsequently ignored by logcheck.

Since there may no longer be any record of that conffile[*], a list of
past MD5 checksums must be provided in $OLD_CONFFILES.  More than one
checksum is allowed per conffile, as the package may have been upgraded
past the last modification, leaving the conffile in an older state.

Unlike with rm_conffile(), we do bother comparing package versions here,
as the call to dpkg-query --search can be costly.  One single version
check should be enough, as we don't plan (with any luck) to add any
files to that list.

[*] Actually, dpkg will keep a record of that conffile up until the
package is purged, or another package adopts it.