Allow filenames with spaces in logcheck.logfiles (closes: #319169)
This replaces the "for VAR in CMD" construct with "CMD | while read VAR"
when parsing logcheck.logfiles.  This allows for logcheck.logfiles to
contain filenames with embedded spaces.  (Leading and trailing spaces
are not retained by read.)

(Note that while there are a couple other "for VAR in CMD" in logcheck,
they all work on rules filenames, which are already filtered by
run-parts and cannot contain spaces.  This is the only one worth
changing.)